feat(Media): add verification on input + modif wording

+++ www/Controller/Media.class.php
@@ -5,7 +5,14 @@ use App\Core\View;
 
 class Media{
     public function index(){
-        if(!empty($_FILES)){ //récupère le fichier
+        $configFormErrors = [];
+
+        if(isset($_POST["submit"])){
+
+        if(empty($_FILES['photo']['name'])){ //récupère le fichier
+            $configFormErrors[] = "Veuillez séléctionner une image pour toute action.";
+        }
+        if(empty($configFormErrors)){
             $target_dir = __DIR__."/../uploads"; //défini le path de notre dossier upload
             if (!is_dir($target_dir)) { //si upload n'existe pas
                 mkdir($target_dir, 0777);
@@ -16,13 +23,28 @@ class Media{
             if(in_array($file_extension,$extension_allow)){//si extension est prise en charge
                 $temp_file = $_FILES['photo']['tmp_name'];  
                 copy($temp_file, $target_dir."/".$file); //copie l'image dans upload
+            }else{
+                $configFormErrors[] = "Les extensions autorisés sont : .jpg, .png, .jpeg";
+            }
+        }
+    }
+        
+
+        if(isset($_POST["delete"])){
+            if(empty($_POST["name-img"])){
+                $configFormErrors[] = "Veuillez choisir une image à supprimer";
+            }
+            if(empty($configFormErrors)){
+                $target_dir = __DIR__."/../uploads/".$_POST["name-img"];
+                if(file_exists($target_dir)){
+                    unlink($target_dir);
+                }else{
+                    $configFormErrors[] = "Ce fichier n'existe pas.";
+            }
             }
         }
         $v = new View("Page/Media","Back");
-        if(isset($_POST["delete"])){
-            $target_dir = __DIR__."/../uploads/".$_POST["name-img"];
-            unlink($target_dir);
-        }
+        $v->assign("configFormErrors", $configFormErrors??[]);
 
     }
 }
